create useNotes hook
